Fixed issue with order > 28
There were multiple ways of specifying how to keep and index and how to remove it, for the cases of 2 or 3 boundaries being resolved, I didn't take the unique values, which led to duplicates.

Simplified the code, which is more uniformly written and much more readable.

Need to re-run these calculations to fix the calculated order for each boundary