Manually adding in T2E fusion in `manually-resolved` breakpoints
We know from prior evidence that this fusion exists in these samples, and the contact matrices display a deletion at these sites.
It appears that these are not picked up due to the thresholding of breakfinder.
To better discuss recurrent SVs and the complex chains they are involved in, I'm adding these manually using the loci identified in PCa51852 (where it's actually detected by breakfinder).