Adding SV comparison between WGS and Hi-C SVs
They don't seem to be in good correspondence between each other, honestly.
They look similar in circos plots and some of the major events, but they don't appear to be picking up many of the same events.
Anecdotally, it seems like the WGS data more easily detects small rearrangements, which is no surprise