Updated plots for TAD calls and boundary persistences.
This set of calls is somewhat suspect for TAD calls, since they seem to
be dependent on the sequencing depth.

Using these raw TAD calls (even after aggregating and filtering) leads
to major differences between tumours and benigns.
Sequencing depth needs to be accounted for.
So while these scripts and plots are updated, the results are not
necessarily useful in the paper itself.