Updated prostate DMRs within CTCF binding site enrichment plots
Conclusion of not looking at hypermethylation globally, still stands.
There is an "enrichment" of hyper-DMRs within CTCF binding sites compared to a global background, but this isn't uniquely sensitive in just observed areas.
This local sensitivity is, however, observed, for the demethylated sites.
Which may indicated looking at a loss of methylation and stronger insulating effect may be more reasonable